Consolidate review tests into test_vllm_to_hf_conversion.py
